Move the store_array_in_hdf5 function to the common functionality
