+++ nac/__init__.py
@@ -4,7 +4,8 @@ from .basisSet import (createNormalizedCGFs, createUniqueCGF,
 
 from .common import (AtomBasisData, AtomBasisKey, AtomData, AtomXYZ,
                      CGF, InfoMO, InputKey, MO, change_mol_units, getmass,
-                     retrieve_hdf5_data, search_data_in_hdf5, triang2mtx)
+                     retrieve_hdf5_data, search_data_in_hdf5,
+                     store_arrays_in_hdf5, triang2mtx)
 
 from .integrals import (calcMtxMultipoleP, calcMtxOverlapP, calc_transf_matrix,
                         calculate_couplings_levine, general_multipole_matrix,
@@ -30,5 +31,5 @@ __all__ = ['AtomBasisData', 'AtomBasisKey', 'AtomData', 'AtomXYZ', 'CGF',
            'getmass', 'initialize', 'lazy_couplings',
            'photoExcitationRate', 'prepare_cp2k_settings',
            'prepare_job_cp2k', 'retrieve_hdf5_data', 'search_data_in_hdf5',
-           'split_file_geometries',
+           'split_file_geometries', 'store_arrays_in_hdf5',
            'store_transf_matrix', 'triang2mtx', 'write_hamiltonians']
